+++ components/button/SelfieButton.tsx
@@ -1,16 +1,15 @@
+import { useCustomToast } from "@/hooks/use-custom-toast";
 import { useRunStore } from "@/store/useRunStore";
 import * as ImagePicker from "expo-image-picker";
 import { useRouter } from "expo-router";
 import React from "react";
 import { Alert, Text } from "react-native";
 import { Button } from "../ui/button";
-import { Divider } from "../ui/divider";
-import { CheckCircleIcon, Icon } from "../ui/icon";
-import { Toast, ToastTitle, useToast } from "../ui/toast";
+import { CheckCircleIcon } from "../ui/icon";
 
 const SelfieButton = () => {
   const router = useRouter();
-  const toast = useToast();
+  const { showToast } = useCustomToast();
 
   const takeSelfie = async () => {
     const { status } = await ImagePicker.requestCameraPermissionsAsync();
@@ -26,25 +25,7 @@ const SelfieButton = () => {
       quality: 1,
     });
     if (!result.canceled) {
-      toast.show({
-        placement: "top",
-        render: ({ id }) => {
-          const toastId = "toast-" + id;
-          return (
-            <Toast
-              nativeID={toastId}
-              className="flex-row items-center gap-4 bg-primary-300 px-5 py-3 shadow-soft-1"
-            >
-              <Icon as={CheckCircleIcon} size="xl" className="text-green-500" />
-              <Divider
-                orientation="vertical"
-                className="h-[30px] bg-outline-200"
-              />
-              <ToastTitle size="sm">사진이 저장되었습니다!</ToastTitle>
-            </Toast>
-          );
-        },
-      });
+      showToast({ message: "인증샷이 저장되었습니다!", icon: CheckCircleIcon });
       useRunStore.getState().addRunData({ selfie: result.assets[0].uri });
       router.push("/running/selfie");
     }
